Added growlGroup & growlAggregate settings.
If growlGroup is set to false, then the results are not growled for groups
If growlAggregate is set to false, the the results are not growled for the aggregate.

Both settings are true by default.

Justification: libnotify notifications can be quite slow to disappear - notify-osd ignores the timeout parameter. So growling each test isn't useful. The aggregate is fine